requirements.txt
BeautifulSoup==3.2.1
beautifulsoup4==4.3.2
cssselect==0.9.1
datrie==0.7
funcsigs==0.4
goose-extractor==1.0.25
jieba==0.37
jsonpickle==0.7.1
libextract==0.0.12
lxml==3.4.0
mock==1.3.0
nltk==3.0.5
nose==1.3.7
numpy==1.9.2
pandas==0.17.0
pbr==1.8.1
Pillow==2.4.0
pluggy==0.3.1
progress==1.2
py==1.4.30
pytest==2.8.2
python-dateutil==2.4.2
pytz==2014.9
PyYAML==3.11
redis==2.10.1
requests==2.3.0
scipy==0.15.1
six==1.10.0
statscounter==0.0.0
tox==2.1.1
virtualenv==13.1.2
wheel==0.24.0
Whoosh==2.6.0
